+++ finagle-stats-core/src/main/scala/com/twitter/finagle/stats/logOnShutdown.scala
@@ -1,0 +1,6 @@
+package com.twitter.finagle.stats
+
+import com.twitter.app.GlobalFlag
+
+object logOnShutdown
+    extends GlobalFlag[Boolean](false, "Enable behavior to dump metrics via logger on shutdown")